Check for versions of gtk that are too new for gattrib.
Since gattrib currently doesn't build with the most recent version of
gtk, add a check for max allowed gtk version.  If the installed gtk
is too recent then error out and suggest the use of --disable-gattrib.

Implement --disable-gattrib to go along with the above check.  Long
term we need to fix gattrib, but that is more involved and this allows
the rest of gEDA/gaf to still be build on the newest of systems.