gattrib: Remove menu entry for unimplemented "File->Open" feature
Popping up a dialog for unimplemented features stops the user being
able to identify what parts of the program are supposed to work.

Due to internal data-structures not being cleared out, opening new
files is not a trivial feature to implement cleanly.

Remove the non-working menu option for now.

Users must select the files they wish to work with as command line
arguments, or in the file-chooser dialog offered when the program
starts.