Modify the drawing of the grips to set a hard limit on the screen size
This patch changes the size of the grips in gschem.  In the past, the
grip size was always directly related to the zoom factor (you would
get really large grips when zoomed in a lot).  Now the grip sizes are a
fixed size at certain zoom factors (grip sizes are reasonably sized when
zoomed in a lot).  This patch also renders the grips on top of endpoint
cues so that you can see the grip rectangle when you are zoomed in a lot
(there might be a better way of doing this, but this works for now).

Signed-off-by: Ales Hvezda <ahvezda@seul.org>