Incremented package version and date to 1.5.4 and 20090830
gEDA/gaf 1.5.3 had some critical bugs in it, specifically:
d388f5370c4b264df90e79a412968589706402c4
4a4218692e089ef40e17acf045cb94849e8b6b3c
and a nice bug fix:
b1f7256890fe8ff48ae077d1fa3cdfcf29ccdc0f

gEDA/gaf v1.5.3 was never widely distributed, so it was withdrawn before
it spread.  One of the bugs dealt with the so versioning on libgeda (so
did not get any version), so I am not going to change the so version at
this point in time (stays at 37:0:0).