gattrib: Remove bit rotten GtkItemEntry widget implementation
This fixes gattrib with Gtk > 2.16.x

The old gtkitementry_2_2.c code subclassed GtkEntry, and (ab)used its
private implementation. The implementation of GtkEntry changed in 2.18 to
allow a model / view architecture, breaking this code (and thus gattrib).

Since we didn't use any of the extra features the GtkItemEntry subclass
provided, we can simply drop this code and use plain GtkEntry widgets in
gattrib.

Remove the configure test which requires gattrib to be disabled when
building against later gtk versions.