Updated the wiki docs from the online version (in prep for the next release)
(Un)fortunately, the previous mechanism for updating the wiki offline
docs from the online version is completely broken now (due to a segfault
in an external tool), so a new mechanism was used to download the files.
The good news is that the number of files and directories that need to
be downloaded has been reduced considerably.  This is the first cut
of the wiki sync.