Don't immediately redraw rubberband objects after placing
This allows the user to see the placed (and selected) object
without it being overwriten by any object still attached to the
mouse cursor. (This made it appear that the object had not been
placed).

The rubberband objects attached to the cursor are re-drawn as soon
as the mouse moves.