Set the LC_NUMERIC locale to "C" rather than "POSIX"
This value is supported on Win32 platforms, whereas "POSIX" doesn't
appear to have any effect.

This is required in order to get correct postscript output in locales
where "," is used as the decimal point separator. It also affects the
font strings passed to Pango, causing broken text rendering in gschem.

Thanks to Cesar for testing this change indeed fixes the issue.

Tested-by: Cesar Strauss <cestrauss@gmail.com>