gschem: Set font description explictly, rather than by parsing a string
This makes the font description string robust against numeric locale
settings, which could cause problems if "," gets used as the decimal
separator in the requested font size.

This issue caused text rendering to be broken in gschem on Win32 for
locales where "," was used as the decimal specifier. The underlying
bug (failing to override the LC_NUMERIC settings) is now fixed,
although it is just as well to be as robust as possible here.

Thanks to Cesar and Bert for testing this fix.

Tested-by: Cesar Strauss <cestrauss@gmail.com>
Tested-by: Bert Timmerman <bert.timmerman@xs4all.nl>