gschem: Improve the text edit dialog's alignment menu
This patch alters the way the alignment for text is set in the text edit
dialogue. I've always felt that the linear list of options doesn't work
very well and slows down my work flow quite a bit.

The GtkOptionsMenu, which is now deprecated, has been replaced with a
GtkComboBox. The alignment options are laid out in a table allowing the
correct option to be selected much quicker.