libgeda: Link tile system contents to page contents.
Instead of adding objects to tile system on object creation and
removing them on object deletion, add and remove objects from tile
system at the same time as they are added to/removed from a page.

This makes more sense, since each page has a separate tile structure.
And has the added advantage of getting rid of a place where the
ADDING_SEL flag is needed!