Kill ADDING_SEL flag with FIRE!
The TOPLEVEL.ADDING_SEL flag has long preyed upon intrepid programmers
venturing into the labyrinthine and treacherous depths of the gEDA
source code, introducing strange and hard-to-trace logic errors and
segmentation faults to hinder them on their various quests to fix bugs
and add features.

But no more! The beast lies slain! No more will developers need to
perform the dance of "flag save and restore" to placate the ravening
beast! No more will they mutter strange incantations for hours as they
seek to determine whether the flag should be set or unset!



In serious terms: the flag was only used for controlling whether or
not objects should be added to the tile system.  By ensuring that
objects are only added to or removed from the tile system when added
to or removed from the tile's corresponding page, it's then possible
to replace the flag check by checking whether the OBJECT is currently
part of a PAGE.