gschem: Cancel box grip operation if height is collapsed to zero
The cancel worked for the case where width was collapsed to zero, but
not height. Seems to be due to 1) To a copy-paste error meaning the
box height was always read as zero, and 2) A logic error in that either
a zero width, OR zero height box should abort the grip opeation.