gschem: Make sure we invalidate the rubberbanded object before cancelling
This showed up as a rendering artefact left behind after grip-resizing
objects to an invalid size, then letting go of the grip. The grip
operation is cancelled, and the original object re-drawn - but it is
possible that the rubber-banded view of the object just before
cancelling the action covers an area of the screen not touched by the
original object.