Use change notification to redraw objects when connectivity changes
Emit change notifications in the connectivity tracking functions
s_conn_remove_other() and s_conn_update_line_object(). This frees
gschem from the somewhat complex task of tracking which objects were
connected to a given object before and after a particular set of
operations.

This change leads to a significant simplification of many operations,
and removes quantities of code from:

  o_pin_end()
  o_net_end()
  o_move_end()
  o_move_end_rubberband()
  o_move_end_lowlevel()
  o_mirror_world_update()
  o_rotate_world_update()
  o_grips_end_bus()
  o_grips_end_pin()
  o_grips_end_net()
  o_delete()