libgeda: Add mechanism to call hook function when an object changes.
Introduces two new functions wihin libgeda, o_emit_pre_change_notify()
and o_emit_change_notify(), intended to be called prior to and after
an object being modified, respectively. This mechanism is primarily
intended as a means for libgeda to trigger on-screen updates when
objects are changed, and as such - at present, only a single pair of
call-backs can be registered (per TOPLEVEL). These call-backs are
registered via the new public function o_set_change_notify_funcs().

The o_emit_pre_change_notify() function gives the GUI an opportunity to
queue a redraw for the screen region where an object used to reside,
prior to operations which change its bounding box, and similarly, the
o_emit_change_notify() allows the GUI to trigger a redraw of the
modified object.