libgeda: Process updates prior to removing an object from the page.
Renames the private function object_removed to pre_object_removed,
and moves its calls prior to removing the object from the page.

For the existing  updates, the order doesn't matter, however other
things we might like to do per-object when removing them from the
page, such as emitting notifications of what we're about to do,
must be done before the object is actually removed.