gschem: Hookup change notify functions to invalidate on-screen objects
When change pre-change and change notifications are emitted on objects,
we need to invalidate their on-screen representation.

The pre-change notification is used to invalidate the area where an
object is currently drawn, which is expected (usually) to be followed
by a change notification - so the new object bounds are invalidated
to be repainted.

Certain operations might skip one of these events. For cases where the
bounding box of an object is known to remain constant, the "pre-change"
notification could potentially be omitted. Adding an object to the page
will (obviously) only be able to generate a "change" notification, and
removing an object will only generate a "pre-change" notification.