gschem: Remove o_invalidate_glist() call from i_callback_edit_(un)embed
This invalidate calls are not required, as any relevant changes will
cause change notifications from within libgeda to trigger redraws of
specific areas of the page.

NB: Changing the embedded status of an object doesn't actually cause a
change notification at present, since it should not affect the how the
objects are rendered.