libgeda: Add weak refs to OBJECT/PAGE/TOPLEVEL.
Weak references are useful when you want to keep an eye on an object
but not necessarily hang on to it when it comes to its useful life.

This patch adds an internal libgeda API for using weak references.
There are two types of weak references implemented by this patch:

* full-powered weak references, where an arbitrary callback function
  can be registered on an object to be notified when an object is
  destroyed

* weak pointers, which are set to NULL when the object is destroyed.

This functionality will hopefully allow several pieces of housekeeping
code in libgeda to be simplified.

The API is designed to be as close as possible to the GObject weak ref
API. See e.g. g_object_weak_ref().