Move STRETCH structure and s_stretch.c from libgeda to gschem
This structure and related functions deal with rubberbanding nets during
move operations. They are only used from within gschem, although there
is of course setup and tear-down in libgeda. Move this all to gschem.

This isn't a 1-1 move, since various headers and structures needed to
be fixed up in the process. The stretch_list now lives in the
GSCHEM_TOPLEVEL, not the PAGE, matching the fact that we don't allow
action state to persist when a GSCHEM_TOPLEVEL changes active PAGE.

A free of the stretch list is removed from s_page_delete(), but no
equivalent is re-added in gschem. Action state is torn down when an
action is cancelled, which is done before deleting any GSCHEM_TOPLEVEL.