libgeda: Add o_box_modify_all().
o_box_modify() allows one corner of a box to be moved.  Unfortunately,
it swaps the internal coordinates of the BOX structure around in a way
that could potentially cause unexpected behaviour.  For instance,
suppose you wish to modify a box as shown below:

   ^                          ^
   |                          |
  4|                         4|       +-----+
   |                          |       |     |
  3|                         3|       |     |
   |                          |       |     |
  2| +--+                    2|       +-----+
   | |  |                     |
  1| |  |                    1|
   | |  |                     |
  0| +--+                    0|
   |                          |
   +--------------->          +--------------->
     0  1  2  3  4              0  1  2  3  4

   (A) Original box           (B) Desired box

Calling o_box_modify() as follows will not work as expected:

  o_box_modify (toplevel, box, 2, 4, BOX_UPPER_LEFT);
  o_box_modify (toplevel, box, 4, 2, BOX_LOWER_RIGHT);

It will instead result in the transformations below:

   ^                          ^
   |                          |
  4|    +--+                 4|    +--------+
   |    |  |                  |    |        |
  3|    |  |                 3|    |        |
   |    |  |                  |    |        |
  2|    |  |                 2|    +--------+
   |    |  |                  |
  1|    |  |                 1|
   |    |  |                  |
  0|    +--+                 0|
   |                          |
   +--------------->          +--------------->
     0  1  2  3  4              0  1  2  3  4

   (C) After first call       (D) After second call

This is due to the fact o_box_modify() behaves as though the corner
specified by the whichone argument had been dragged.

This patch adds the o_box_modify_all() function, which allows all four
corners of the box to be modified at once, avoiding the problems
inherent in repeated o_box_modify() calls.