gschem: Remove global_window_current.
Instead, use the GSCHEM_TOPLEVEL fluid (accessed via
g_current_toplevel()). Although this does indeed remove one magic
global variable in favour of another, the fluid has the advantage of
coping correctly with non-local exits, and is more clearly only for
the benefit of procedures that are called via Scheme.