gschem: Add a fluid for current GSCHEM_TOPLEVEL.
Adds a smob that represents a gschem window structure, and a Guile
fluid that tracks the active window and that can be accessed from
Scheme code.  The functions which update the fluid also ensure that
the libgeda TOPLEVEL fluid is updated in lockstep.  This ensures that
Scheme code that is not gschem-aware will run correctly in gschem, and
will allow the global_window_current variable to be removed.