scheme-api: Pin objects.
Re-uses the line modification routines for working on pins.  Note that
the libgeda C API provides a flag for choosing which end of a pin
should be connectable, but the Scheme API hides this flag, and from
the point of view of Scheme code the connectable end of a pin is
always the start.