scheme-api: Complex objects.
Support for working with complex objects.  Because Guile supports
complex numbers natively, using the word "complex" in function names
etc. is confusing.  In the public Scheme API the word "component" is
used instead.

This patch implements support for inspecting and modifying components,
as well as for creating new, empty embedded components.  It does not
include support for creating components from files or from the
component library.