libgeda: Define non-NULL default for select_func.
The code for loading/saving complex objects uses the nullity of
OBJECT.sel_func to determine whether or not a complex object should be
flagged as selectable or not.  So that libgeda applications that do
not define their own select_func can safely load, modify and save
schematic files, it's necessary to have a non-NULL default
select_func.