scheme-api: Attributes.
Adds support for attributes.  In order to make the API safe to use
without memory corruption, there are significant restrictions on the
context in which attribute attachments can be made.  Most importantly,
both target and attribute must be part of the same page, or of the
same component, and objects which have attribute attachments cannot be
removed from a page or component's contents.