scheme-api: Basic smob definitions & TOPLEVEL fluid.
Adds a new smob type for gEDA objects (TOPLEVEL, PAGE and OBJECT, at
least initially), along with C functions for creating smobs.  Weak
references are used to ensure that smobs pointing to dead objects are
invalidated.

Additionally, defines a fluid which contains the current TOPLEVEL.
This is used by the Scheme API to obtain the TOPLEVEL needed for
calling much of the libgeda API.