scheme-api: Closing pages.
Since it's possible to create new pages from the Scheme API, it's also
useful to be able to close them.  This patch adds the close-page!
function to the (geda page) module.

It's necessary for gschem to provide a slightly different
implementation of close-page! which allows for the GUI to be updated
and for the user to be prompted to save changes.  To do this, gschem
uses Guile module reflection to change the binding of the close-page!
function during startup.