Remove most legacy libgeda Scheme types & functions.
Removes most of the legacy Scheme types, functions and definitions
from libgeda, replacing them with compatible equivalents based on the
new Scheme API.  The remaining Scheme functions are those for working
with the component and source libraries, the various rc functions, and
the eval-* functions, all of which need to be revisited at a later
date.

Also removes set-attribute-value! from gschem and reimplements in
Scheme, since it was sole user of g_set_attrib_value_internal() and it
was easier to replace than rewrite.