Remove DONT_REDRAW flag.
Because we now use the GTK/Cairo renderer for gschem, we use the GTK
model for redrawing. In this model, objects are queued for redrawing
by invalidating them, and at some point later all invalidated regions
of the viewable area are drawn together.

Due to this mechanism, the TOPLEVEL.DONT_REDRAW flag is redundant, as
it as always set and then restored without visiting the GTK main loop,
and thus no drawing would occur in that context anyway.

This patch removes the now-useless TOPLEVEL.DONT_REDRAW flag entirely.