gnetlist: Handle hierarchy refdes correctly in partslist*. [3044478]
When producing a part list with the partslist[1-3] backend the refdes
values output to the file do not contain the prepended refdeses of the
parent component.

E.g.

Master schematic with symbol 'S1' (representing sub-schematic) and a
resistor 'R1' Sub-schematic containing resistor 'R1'

The geda backend produces a netlist containing two resistors 'R1' and
'S1/R1', they're unique and everything is fine and dandy.  The
partslist[1-3] backends contain two components with the refdes 'R1'
which then causes problems when processing this file further.

This is caused by the partslist[1-3] backends adding the value of the
refdes attribute to the file rather than the value returned by the
'packages' function.

The supplied patch alters the backends to use the value from the
packages function directly, rather than looking up the refdes
attribute of the component.