libgeda: Always add gEDA default Scheme directory to search path.
Make sure that the default directory for gEDA Scheme files is added to
the Guile %load-path during libgeda initialisation.  This ensures that
even if there is a problem loading system-gafrc (for some reason),
gEDA Scheme code can still be found.

Remove the current `scheme-directory' directive from system-gafrc, and
replace it with a commented-out example of its use.