gnetlist: NoConnection nets with more than 1 pin are an error in drc2 backend.
Adding the NoConnection DRC attribute to a net causes the drc2 backend
to ignore it when checking for unconnected nets. However, drc2
currently doesn't enforce that the net is actually unconnected. It's
possible to mark two pins as NoConnection and then accidentally wire
them together. This patch to gnet-drc2.scm checks for that case.

Reviewed-by: Peter TB Brett <peter@peter-b.co.uk>
Affects-bug: lp-698743