gnetlist: Add convenience library to help backends parse `-O' options.
Based on a suggestion by Patrick Bernaud that it would be useful for
gnetlist backends to be able to accept more complex options on the
gnetlist command line, this patch adds the (gnetlist backend-getopt)
Scheme module to gnetlist.

This provides convenience functions that allow a backend to easily
handle several different types of `-O' backend option: flag options,
options with arguments, and options with optional arguments.