gsch2pcb: Don't pass invalid function names to m4.
This fixes a long-standing bug with gsch2pcb.  gsch2pcb expects
`footprint' attributes to be of the form:

  footprint=NAME ARG1 ARG2

This gets transformed to:

  PKG_NAME (`NAME-ARG1-ARG2',`refdes',`value',`ARG1',`ARG2')

for macro-expansion by M4.  For this to work, it requires that NAME
matches the regular expression ^[A-Za-z0-9_]$ so that PKG_NAME is a
valid M4 function name.

It is quite common for users (especially those using John Luciani's
footprint library) to define and wish to use footprint names which
contain hyphens '-'.  Since this character is not valid in M4 function
names, this practice causes non-obvious breakage to occur.

This patch stops gsch2pcb from ever attempting to macro-expand a
footprint NAME which is obviously not valid as part of an M4 function
name.

Closes-bug: lp-698806