gschem: Make dragging selected objects obey the same hit test as selecting
This makes dragging a group of selected objects more consistent with the
action required to select the objects in the first place, most notably
when dealing objects whos bounding boxes don't fit tightly to the
selectable geometry - e.g. lines and curves.

In order to start a drag, the user must hit within some distance of some
selected geometry. This makes it easier to change selection to another
object with overlapping bounding box if the user did not intend to initiate
a drag.