gsch2pcb: Follow g_spawn_sync() documentation and pass required flags
The documentation for g_spawn_sync() tells us that we must pass the
flags G_SPAWN_STDOUT_TO_DEV_NULL and G_SPAWN_STDERR_TO_DEV_NULL if we
pass NULL for the stdout and stderr returns from the command.

These flags cause g_spawn_sync() to redirect stdout and stderr to
/dev/null rather than assigning pipes to them.

No known bug is fixed by this, but it is possible that without these
flags, either the stdout or stderr pipes' buffer could become full if
the command executing emits enough data. This might lead to both
processes stalling.

Whilst I was touching this code, I added some comments as to the
meaning of the various NULL arguments we pass to g_spawn_sync().