gschem: Make the user click the first point when adding nets or buses
This allows "magnetic net" mode to be used for the start-point of a net
as well as its end-point, whilst still being able to initiate net drawing
with the "n" key.

I have updated buses as well in an effort not to have them diverge yet
further in their behaviour.