gschem: recent file list checks file extensions
In the recently used files both symbol and schematic files are shown.
Change the filtering method, so recently used files are chosen not only
by MIME type, but also by the file extension as well.

Rationale: if gEDA is installed in a location other than default (say in
users home dir), the MIME type databases will not be updated and list
of recently used files will always be empty.