auto-uref: smarter calculation of first refes on page
Introduces the "page-offset" concept, where it is assumed
refdeses on each page should start at multiple of page-offset.
So for page-offset = 100 and minimum refdes used on page = 224
the hook will assign next refdes equal to 201.

To enable this feature use (auto-uref-set-page-offset <number>).
Default value of page-offset is 0 which disables the feature.

Currently minimum refdes is calculated separately for each prefix.