libgeda: Fix crash on startup on Windows.
If the GEDADATA environment variable was not set on Windows, libgeda
was unable to determine the gEDA system data directory. It would then
attempt to make a Scheme string from the resulting NULL pointer,
causing a crash.

This patch alters the behaviour on Windows. If GEDADATA is not set,
libgeda now guesses that the system data directory is the current
working directory.

Unfortunately, this will still prevent most gEDA programs from working
due to not being able to locate data files, but it will prevent an
uninformative segfault on startup.

Affects-bug: lp-795545