scheme-api: Add add-attrib! function.
Add convenience Scheme API function in gschem to shadow the
o_attrib_add_attrib() C function.  Although most of the behaviour of
this function could be provided in pure Scheme, it would be somewhat
annoying to have two lots of code which do the same thing.  Also, the
slotting mess is still in C only.