build-sys: Test for scm_{to,from}_utf8_string().
In gEDA, we use UTF-8 strings, but Guile doesn't.  The behaviour of
scm_{to,from}_locale_string() functions depends on the current locale,
and may not be the right thing if the user has set a non-UTF-8 locale.
This patch adds a configure test for the availability of the
scm_{to,from}_utf8_string() functions, which will do the right thing
with strings passed from gEDA to Guile independently of the locale.