Updated the problematic name attributes in some symbols
In some gnetlist backends the name=? attributes cause problems because
they look like valid attribute values. eg file=? may be mistaken for a
valid filepath, and value=? for an actual value. Changing them to
name=unknown is one way to solve this problem in the interim. It's now
safe to remove the special checks for these "?" like the one in
spice-sdb. This patch updates these problematic name=? attributes in
some symbols to name=unknown, and removes FIXME code snipit that
checked for the ? value in spice-sdb.

It also updates gnetlist tests to reflect these changes. Turns out
that the test input file TwoStageAmp.sch had a ? in it, which of couse
caused problems.  Then when it was removed vams's output changed so it
had to be fixed, which wasn't a very clever fix but seems to work.

Closes-bug: lp-698741
Reviewed-by: Peter TB Brett <peter@peter-b.co.uk>