scheme-api: Make gschem's overridden close-page! never ask.
The original version of the overridden version of close-page! loaded
by gschem into the (geda core page) asked the user to confirm closing
of dirtied pages.  It turns out that this was a really bad idea,
because it made the close-page! function impossible to document
clearly.  This patch makes the behaviour of gschem's close-page! match
the core library version, i.e. close immediately without question.