scheme-api: Run gtk-update-icon-cache when uninstalling.
If `gtk-update-icon-cache' is not run during `make uninstall', then a
stale `icon-theme.cache' is left in the install directory.  This
breaks distcheck's uninstall check.

This problem was previously masked due to the fact that the uninstall
check silently allows exactly one leftover file -- and now that we
install Info manuals, this allowance is being used up by the
`${infodir}/dir' file generated by `install-info'.