Improved error handling in gsch2pcb.
gsch2pcb: The return code from g_spawn_sync is checked when executing gnetlist.
If g_spawn_sync is successful, the exit code from gnetlist is checked, and
anything but a non-zero exit will be flagged as an error, and standard error
from gnetlist will be forwarded to the terminal. Notice that gnetlist may have
a successful exit code, even when errors are encountered, see bug lp-700448.

Affects-bug: lp-806686