Revert "gschem: Make the user click the first point when adding nets or buses"
This reverts commit 2197952617d42ccbe2355747ef48668c63ece30b.

It turns out this behaviour was unpopular, and until I (or someone) has
time to make the behaviour configurable, it is better to revert this.

Perhaps we may be able to come up with another way of allowing magnetic
net mode for the start-point of a net, whilst retaining the immediate
action of the shortcut key.

(Revert tweaked a little to avoid re-adding whitespace errors)