gnetlist: Regenerate `bom' and `bom2' test results.
In commit 76bc345ee043, the `bom' and `bom2' backends were modified to
be more informative about attribute list specifications.  As a
side-effect of this work, the backend options for these backends are
now much more strictly validated.  Tests which pass `-O include_mode'
or `-O sort_mode' to gnetlist therefore now fail for the BOM backends,
and this patch updates the golden files appropriately.

Affects-bug: lp-810202
Reported-by: Dan McMahill <dan@mcmahill.net>