gschem: Key combinations as first-class objects.
Adds a new Scheme type to gschem that represents a key combination
(e.g. Ctrl+Shift+A).  The interface for working with them simply wraps
the GTK+ accelerator handling API, which means it's easy to convert
the key combination smobs either into a precise representation for
storing or into a properly-translated "pretty" representation for
display.