gschem: Rewrite gschemdoc in Scheme.
This completely removes the "gschemdoc" helper tool, and reimplements
almost all its functionality in Scheme, using the "show-uri" function
added in commit b4452ce0b1d4.  This permits gschem's documentation and
component documentation to be found and shown in an appropriate viewer
on all platforms, including Windows.  I have verified that this works
correctly on Linux, Cygwin and Windows XP.

There is a drawback, however, in that "locate" is no longer used to
search for candidate files -- only the gEDA documentation directory
and the working directory are checked.  In future, this functionality
should probably be restored using desktop search.  There may also be
problems displaying documentation if gEDA data or documentation were
installed to non-standard directories, due to the slightly inelegant
methods used to detect the gEDA documentation directory relocatably.

The component documentation search behaviour can be altered by
redefining the 'hierarchy-documentation' function in a gschemrc file.