gschem: Add x_show_uri().
This function launches a URI in the appropriate application for
handling it, as determined by the platform's filetype association
settings.

GTK+ 2.16.0 and later provide a the gtk_show_uri() function, which
attempts to use GIO to launch the default application for a URI.
Unfortunately, it doesn't work properly on Windows with HTTP and FTP
URLs, does evil things with FUSE when passed some URIs, and insists on
GNOME applications even when you're not running GNOME.  Unfortunately,
it is also currently considered the Right Way of doing things, and has
the advantage of providing detailed error messages when things go
wrong.

With this patch, x_show_url() defaults to using gtk_show_url() on
Linux, but not on other host platforms.  The --enable-gio and
--disable-gio arguments to configure can be used to override this
behaviour.

If GIO is not enabled, the default behaviour differs depending on the
platform:

- On native Windows, the ShellExecute() API call is used by default.

- On other platforms, a helper command is used.  If the URI_VIEWER
  variable was not set during configure, the command used is:

  * `cygstart' on Cygwin
  * `open' on Mac OS X
  * `xdg-open' otherwise.

Using a helper command can be forced on Windows by setting URI_VIEWER.

For example, to disable GIO and select an alternative URI launcher
command:

  ./configure URI_VIEWER=my-open --disable-gio