General fixes for pictures, picture rendering and picture operations.
Streamlined the API for picture objects to make it harder to leave a
picture object in an inconsistent state.  This is a prerequisite patch
to providing a Scheme API for working with picture objects.

This patch also makes primitive transformations of pictures actually
work, and fixes rendering of pictures in gschem, because the changes
weren't really feasible to separate.