gschem: Fix a crash when cancelling out of an action.
In some cases, the new keybinding code could cause a crash when
cancelling out of an action with the right mouse button.  This was
because there was an call to scm_c_eval_string() which could sometimes
be called in a context where Guile exceptions weren't caught, and
which hadn't been updated to reflect the changes to the keybinding
system.