gschem: Workaround for zero-height text view bug.
This is a temporary workaround for a bug on systems using overlay
scrollbars for GTK+ widgets, where the value editing GtkTextView in
the multi-attribute editing dialog was appearing with zero height.

Obviously, setting the height request to a constant size isn't ideal.
It would be better to set the height request based on the font size,
so that a certain number of lines are displayed.

This should be revisited at a later date.

Affects-bug: lp-876299