libgeda: Fix null dereference when fallback image not loaded.
In commit bccf15c104d6, a bug was introduced whereby libgeda
dereferenced a NULL pointer.  If a picture object's image cannot be
loaded, libgeda attempts to load a fallback image, and on failure,
accesses a GError to retrieve the error message.  Unfortunately, the
GError was not being passed to the image loading function, and was
therefore uninitialised when it was dereferenced to obtain a message.

Closes-bug: lp-904639