gschem: Correctly construct expressions for running hooks.
In commit 71cc3eb5583b, the way hooks were run was changed from
directly calling libguile's scm_run_hook().  Instead, an expression
was constructed as a Scheme list structure, to be evaluated.

Unfortunately, the way that the expression was constructed was wrong,
and led to the Guile evaluator attempting (and failing) to evaluate
gEDA OBJECT smobs.