gattrib: Make "Add new attribute column" actually insert column.
I first saw bug 698608, which seems to be a duplicate of this bug, and
came up with a patch that corrects the behavior and inserts the empty
data column under the right column heading. The problem was that the
header gets alphabetically sorted into the others (after device,
footprint, and value) but the code immediately following that inserts
the new column after the last column, which seems to be making the
assumption that the headers do not get sorted.

My patch adds a new s_string_list function
(s_string_list_find_in_list()) that searches for a given string in the
list and returns the index (or -1, if it's not there), unlike the
existing function (s_string_list_in_list()) (that I copied from but
didn't touch) which just returns 1 as soon as it finds a matching
string, or 0, if it's not there. I then added code in
s_toplevel_add_new_attrib() that uses the new function to determine
where the new column header was placed and insert the new data column
at that index (instead of at the end).

Closes-bug: lp-698575
Reviewed-by: Peter TB Brett <peter@peter-b.co.uk>