gschem: Draw arrows on 'connected' dangling net ends.
Recently, gschem was changed to not draw anything at a net endpoint
when it thinks the net is connected.  This has the drawback that in
some cases the eye can be fooled into thinking that two nets are
connected together when they are not, with the concomitant risk of
badly breaking a design.

This patch alters gschem to draw an arrow in 'junction colour' on
dangling net segment ends when it thinks the net is properly
connected.  This is not an ideal solution, but should go most of the
way towards solving the "invisible errors wrecking my expensive
boards" problem.

A follow-up patch should probably do the same for printed output.

Affects-bug: lp-1014160