libgeda: Use light background colour map for printing by default.
When users enable colour printing, it can be very confusing for them
to suddenly receive an (expensive) printout covered almost entirely in
black ink.  In order to make users less confused -- and for the
protection of those unfortunates still using inkjet colour printers --
this patch changes the default print colormap to the "light" version.

Affects-bug: lp-698611